TINY-12660: Fix flaky webdriver tests (#10824)
